editoast: move conflicts logic to a specific module
